Add auto-login as house on GR pc.
