--- app/src/app/services/auth-guard.service.ts
@@ -1,20 +1,0 @@
-import {Injectable} from '@angular/core';
-import {CanActivate, Router} from '@angular/router';
-import {AuthService} from './auth.service';
-
-@Injectable()
-export class AuthGuardService implements CanActivate {
-    constructor(
-        public auth: AuthService,
-        public router: Router) {
-    }
-
-    canActivate(): boolean {
-        if (!this.auth.isAuthenticated()) {
-            this.auth.logout();
-            this.router.navigate(['login']);
-            return false;
-        }
-        return true;
-    }
-}
