+++ app/src/app/services/easter.service.ts
@@ -13,6 +13,9 @@ export class EasterEggService {
         'No!?',
         'Scheiβe',
         'Kutsjaarsch',
+        'Miele, er is geen betere',
+        'Yesh',
+        'Yaaas',
         'Twaarsch',
         'Jawel \"huisgenoot\"',
         'Shit',
